Try to get a token from the management API before redirecting.
This should work in all situations. When a user has not yet logged into the management API, this will grant them access based on their IP address without having to redirect to the API and back. If they have logged in, the management API should have redirected to Auth Proxy to set a token, which bypasses the token generation process altogether.